refactor: Removal of obsolete `SkillViewModel`
